Remove cycles from Defaults.scala
Classpaths, BuildExtra, RemoteCache and DeprecatedContinuous had cyclic
dependencies with Defaults. This commit factors out the shared code
into helper objects to remove the cyclic dependencies.
