+++ main/src/main/scala/sbt/internal/BuildDef.scala
@@ -19,8 +19,7 @@ trait BuildDef {
   def projectDefinitions(@deprecated("unused", "") baseDirectory: File): Seq[Project] = projects
   def projects: Seq[Project] =
     CompositeProject.expand(ReflectUtilities.allVals[CompositeProject](this).values.toSeq)
-  // TODO: Should we grab the build core settings here or in a plugin?
-  def settings: Seq[Setting[_]] = Defaults.buildCore
+  def settings: Seq[Setting[_]] = GlobalDefaults.settings
   def buildLoaders: Seq[BuildLoader.Components] = Nil
 
   /**
